% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_similarity_models.R
\name{bow_tie}
\alias{bow_tie}
\title{Function to compute bow tie similarity}
\usage{
bow_tie(behave_data, center_value)
}
\arguments{
\item{behave_data}{ordered vector of data you want to compute similarity on}

\item{center_value}{the value around which you want to center your similarity matrix}
}
\value{
n x n similarity matrix
}
\description{
Function to compute bow tie similarity
}
