% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_similarity_models.R
\name{divergence}
\alias{divergence}
\title{Function to compute divergence similarity}
\usage{
divergence(behave_data)
}
\arguments{
\item{behave_data}{ordered vector of data you want to compute similarity on}
}
\value{
n x n similarity matrix
}
\description{
Function to compute divergence similarity
}
