% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_similarity_models.R
\name{fix_diagonal}
\alias{fix_diagonal}
\title{Function to scale your similarity matrix and set diagonal to 1}
\usage{
fix_diagonal(similarity_matrix, index, organize_row)
}
\arguments{
\item{similarity_matrix}{the matrix you get out of running one of the similarity models (i.e., convergence(your_mat))}

\item{index}{the ordered list of subject/participant ids}

\item{organize_row}{set to TRUE if you want to add rownames that correspond to your ids to your matrix}
}
\value{
clean similarity matrix
}
\description{
Function to scale your similarity matrix and set diagonal to 1
}
