% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_similarity_models.R
\name{make_similarity_matrices}
\alias{make_similarity_matrices}
\title{Function to compute all similarity models at once and collect them into a single list}
\usage{
make_similarity_matrices(behave_vec, index, center_value, resultpath, save_out)
}
\arguments{
\item{behave_vec}{ordered vector of data you want to compute similarity on}

\item{index}{the ordered list of subject/participant ids}

\item{center_value}{the value around which you want to center your bow_tie similarity matrix}

\item{resultpath}{the path to the directory where you want your resulting similarity matrices to be stored as .csv files}

\item{save_out}{if set to TRUE, will save out your similarity matrices as .csv files in your resultpath directory}
}
\value{
list containing all four similarity matrices (nearest neighbours, convergence, divergence, and bow-tie)
}
\description{
Function to compute all similarity models at once and collect them into a single list
}
