% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_similarity_models.R
\name{nearest_neighbours}
\alias{nearest_neighbours}
\title{Function to compute nearest neighbor similarity}
\usage{
nearest_neighbours(behave_data)
}
\arguments{
\item{behave_data}{ordered vector of data you want to compute similarity on}
}
\value{
n x n similarity matrix
}
\description{
Function to compute nearest neighbor similarity
}
