% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_similarity_models.R
\name{rank_subxsub_mat}
\alias{rank_subxsub_mat}
\title{Function to reorder and scale N x N+1 matrix by index}
\usage{
rank_subxsub_mat(df, index, df_index_id, scale_matrix)
}
\arguments{
\item{df}{symmetric matrix you wish to re-order}

\item{index}{ordered list of subject/participant ids}

\item{df_index_id}{string value that is the column name of participant ids}

\item{scale_matrix}{set to TRUE if you want to scale your matrix from 0 to 1, FALSE if you don't}
}
\value{
original symmetric matrix ordered by the index order
}
\description{
This function will reorder (and scale to 0->1 if scale_matrix is set to TRUE)
a symmetrical matrix that has an additional column (hence NxN+1) containing
the subject/participant IDs, as they are currently ordered.
This function takes as input an index of subject/participant ids that is already
ordered by some metric (i.e. ascending height). It then re-orders your symmetrical
matrix, removes the ID column returning it to its NxN state, and returns
a re-ordered symmetric matrix
}
