% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_similarity_models.R
\name{scale_matrix}
\alias{scale_matrix}
\title{Function to rescale a matrix between 0 and 1}
\usage{
scale_matrix(matrix)
}
\arguments{
\item{matrix}{symmetrical matrix of numerical values}
}
\value{
the inputted matrix, but rescaled from 0 to (+) 1
}
\description{
This function takes a symmetrical matrix as input and re-scales
it to values ranging from 0 to (+) 1
}
